+++ lib/test_prof/factory_prof/printers/simple.rb
@@ -1,15 +1,19 @@
 # frozen_string_literal: true
 
+require "test_prof/ext/float_duration"
+
 module TestProf::FactoryProf
   module Printers
     module Simple # :nodoc: all
       class << self
+        using TestProf::FloatDuration
         include TestProf::Logging
 
-        def dump(result)
+        def dump(result, start_time:)
           return log(:info, "No factories detected") if result.raw_stats == {}
           msgs = []
 
+          total_run_time = TestProf.now - start_time
           total_count = result.stats.sum { |stat| stat[:total_count] }
           total_top_level_count = result.stats.sum { |stat| stat[:top_level_count] }
           total_time = result.stats.sum { |stat| stat[:top_level_time] }
@@ -21,7 +25,7 @@ module TestProf::FactoryProf
 
                Total: #{total_count}
                Total top-level: #{total_top_level_count}
-               Total time: #{format("%.4f", total_time)}s
+               Total time: #{total_time.duration} (out of #{total_run_time.duration})
                Total uniq factories: #{total_uniq_factories}
 
                  total   top-level     total time      time per call      top-level time               name