+++ spec/integrations/factory_prof_spec.rb
@@ -8,7 +8,7 @@ describe "FactoryProf" do
       expect(output).to include("FactoryProf enabled (simple mode)")
 
       expect(output).to include("Factories usage")
-      expect(output).to match(/Total: 26\n\s+Total top-level: 14\n\s+Total time: \d+\.\d{4}s\n\s+Total uniq factories: 2/)
+      expect(output).to match(/Total: 26\n\s+Total top-level: 14\n\s+Total time: \d{2}+:\d{2}\.\d{3} \(out of \d{2}+:\d{2}\.\d{3}\)\n\s+Total uniq factories: 2/)
       expect(output).to match(/total\s+top-level\s+total time\s+time per call\s+top-level time\s+name/)
       expect(output).to match(/\s+16\s+8\s+(\d+\.\d{4}s\s+){3}user\n\s+10\s+6\s+\s+(\d+\.\d{4}s\s+){3}post/)
     end
@@ -21,6 +21,26 @@ describe "FactoryProf" do
       expect(output).to include("FactoryFlame report generated: ")
 
       expect(File.exist?("tmp/test_prof/factory-flame.html")).to eq true
+    end
+
+    specify "nate printer", :aggregate_failures do
+      output = run_rspec("factory_prof", env: {"FPROF" => "nate_heckler"})
+
+      expect(output).to match(/Time spent in factories: \d{2}+:\d{2}\.\d{3} \([\d.]+% of total time\)/)
+    end
+
+    specify "with nate printer always enabled", :aggregate_failures do
+      output = run_rspec("factory_prof_with_nate")
+
+      expect(output).to match(/Time spent in factories: \d{2}+:\d{2}\.\d{3} \([\d.]+% of total time\)/)
+    end
+
+    specify "with nate printer always enabled and flamegraph profiler", :aggregate_failures do
+      output = run_rspec("factory_prof_with_nate", env: {"FPROF" => "flamegraph"})
+
+      expect(output).to match(/Time spent in factories: \d{2}+:\d{2}\.\d{3} \([\d.]+% of total time\)/)
+      expect(output).to include("FactoryProf enabled (flamegraph mode)")
+      expect(output).to include("FactoryFlame report generated: ")
     end
 
     context "when no fabrication installed" do